requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py[voice]==2.4.0
discord-ext-voice-recv
distro==1.8.0
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai[realtime]==1.58.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pydantic==2.5.1
pydantic_core==2.14.3
requests==2.28.2
sniffio==1.3.0
tqdm==4.65.0
typing_extensions<5
urllib3==1.26.15
yarl==1.8.2
tzdata
pydub